modeling/requirements_predictor.txt
pyspark==3.3.1
psycopg2-binary==2.9.6
numpy==1.23.5
